+++ sample-apps/react-native/dogfood/src/components/ChatWrapper.tsx
@@ -1,12 +1,10 @@
-import React, { PropsWithChildren, useCallback, useMemo } from 'react';
+import React, { PropsWithChildren, useMemo } from 'react';
 import { Chat, OverlayProvider, Streami18n } from 'stream-chat-react-native';
 import { useChatClient } from '../hooks/useChatClient';
 import { AuthenticationProgress } from './AuthenticatingProgress';
 import { StreamChatGenerics } from '../../types';
 import { useAppGlobalStoreValue } from '../contexts/AppContext';
-import { createToken } from '../modules/helpers/createToken';
 import { useStreamChatTheme } from '../hooks/useTheme';
-import { STREAM_API_KEY } from '../../config';
 
 const streami18n = new Streami18n({
   language: 'en',
@@ -26,14 +24,8 @@ export const ChatWrapper = ({ children }: PropsWithChildren<{}>) => {
     [userId, userName, userImageUrl],
   );
 
-  const tokenProvider = useCallback(async () => {
-    return await createToken({ user_id: userId });
-  }, [userId]);
-
   const chatClient = useChatClient({
-    apiKey: STREAM_API_KEY,
     userData: user,
-    tokenProvider: tokenProvider,
   });
   const theme = useStreamChatTheme();